reorganized typing
