+++ react/src/_core/redux/types/index.js
@@ -1,8 +1,10 @@
 // @flow
+import actionStatus, {ActionStatus} from './actionStatus';
+
 export type Action = {
   type: string,
   payload?: any | null,
-  state?: string | null
+  state?: ActionStatus | null
 };
 export type State = Object;
 export type Dispatch = (action: Action | ThunkAction | PromiseAction | Array<Action>) => any;
@@ -10,3 +12,4 @@ export type GetState = () => State;
 export type ThunkAction = (dispatch: Dispatch, getState: GetState) => any;
 export type PromiseAction = Promise<Action>;
 
+export default {actionStatus};
