+++ react/src/_redux/action/action_Archive_Item/action_Archive_Item_Duplicate/index.js
@@ -1,5 +1,5 @@
 // @flow
-import type {Action, Dispatch} from '../../../types';
+import type {Action, Dispatch} from 'src/_core/redux/types';
 
 export const TYPE = 'TYPE_ARCHIVE_ITEM_DUPLICATE';
 export default () => (dispatch: Dispatch) => new Promise((resolve, reject) => {
