+++ react/src/_redux/action/action_Archive_Item/action_Archive_Item_Rename/index.js
@@ -1,5 +1,5 @@
 // @flow
-import type {Action, Dispatch} from '../../../types';
+import type {Action, Dispatch} from 'src/_core/redux/types';
 
 export const TYPE = 'TYPE_ARCHIVE_ITEM_RENAME';
 export default (fileName) => (dispatch: Dispatch) => new Promise(
