+++ react/src/_redux/action/action_Archive_Item/action_Archive_Item_Select/index.js
@@ -1,11 +1,11 @@
 // @flow
-import type {Action, Dispatch} from '../../../types';
+import type {Action, Dispatch} from 'src/_core/redux/types';
 
 export const TYPE = 'TYPE_ARCHIVE_ITEM_SELECT';
-export default (uuid) => (dispatch: Dispatch) => new Promise(
+export default (selectedUuid: string) => (dispatch: Dispatch) => new Promise(
     (resolve, reject) => {
       try {
-        dispatch(({type: TYPE, payload: uuid}: Action));
+        dispatch(({type: TYPE, payload: selectedUuid}: Action));
         resolve();
       } catch (exc) {
         reject(exc);
