+++ react/src/_redux/action/action_Archive_Suite/action_Archive_Suite_Blank/index.js
@@ -1,5 +1,5 @@
 // @flow
-import type {Action, Dispatch} from '../../../types';
+import type {Action, Dispatch} from 'src/_core/redux/types';
 
 export const TYPE = 'TYPE_ARCHIVE_SUITE_BLANK';
 export default () => (dispatch: Dispatch) => new Promise((resolve, reject) => {
