+++ react/src/_redux/action/action_FormTemplate/action_FormTemplate_GetFormsIndex/action_Forms_GetFormsIndex.test.js
@@ -4,7 +4,7 @@ import thunk from 'redux-thunk';
 import ipcRWrapper from 'src/_core/electron/IpcRWrapper';
 import sinon from 'sinon';
 import {ACT_GET_FORMS_INDEX} from 'src/_core/contract/formsBridge';
-import * as actionStatus from 'src/_core/redux/actionStatus';
+import * as actionStatus from 'src/_core/redux/types/actionStatus/index';
 
 describe('action_FormTemplate_GetFormsIndex should ', () => {
