+++ react/src/_redux/action/action_FormTemplate/action_FormTemplate_GetFormsIndex/index.js
@@ -1,6 +1,6 @@
 // @flow
-import type {Action, Dispatch} from '../../../types';
-import * as actionStatus from 'src/_core/redux/actionStatus';
+import type {Action, Dispatch} from 'src/_core/redux/types';
+import * as actionStatus from 'src/_core/redux/types/actionStatus/index';
 import ipcRWrapper from 'src/_core/electron/IpcRWrapper';
 import {ACT_GET_FORMS_INDEX} from 'src/_core/contract/formsBridge';
