+++ react/src/_redux/action/action_Nav/action_Nav_PathChange/index.js
@@ -1,4 +1,4 @@
-import type {Action, Dispatch} from '../../../types';
+import type {Action, Dispatch} from 'src/_core/redux/types';
 
 export const TYPE = 'TYPE_NAV_PATHCHANGE';
 export default (newPath: string) => (dispatch: Dispatch) => new Promise(
